refactor(graphql): move schema/dataloader/transport/pipeline into responsibility-based subfolders
Move schema.ts, dataloader.ts, transport.ts, and input-pipeline.ts
(with their test files) into schema/, dataloader/, transport/, and
pipeline/ subfolders respectively.

- Update all internal import paths to match new file locations
- Add root-level dataloader.ts re-export stub to preserve index.ts
  unchanged (public API signature unaffected)
- All 52 existing tests pass without modification to test logic

Addresses #857
